Benvo data objects

A custom data structure for handling point pattern built environment data. Consists of two types of dataframes (1) subject and (2) subject - BEF with a joining id(s). Whether using the benvo constructor helper function or the raw Benvo class constructor, certain constraints are maintained within the class. Namely,(1) there must be a bef name and component for each bef data frame. (2) Both bef_data and subject data must inherit from class data.frame. (3) Components must be one of c("Distance","Time","Distance-Time")

Slots

subject_data

data.frame containing subject level covariates.

bef_data

list containing subject - bef distance/time data frames

longitudinal

boolean value indicating whether the Benvo holds longitudinal data

bef_names

character vector containing the names of the BEFs

components

character vector containing the Distances and/or Time data measured for each BEF